+++ packages/core/demo/demo14.ts
@@ -1,5 +1,5 @@
-import * as T from "../Effect"
-import * as M from "../Managed"
+import * as T from "../src/next/Effect"
+import * as M from "../src/next/Managed"
 
 export const man = M.onExitFirst_(
   M.fromEffect(
