+++ packages/core/demo/demo6.ts
@@ -1,5 +1,5 @@
-import { pipe } from "../../Function"
-import * as T from "../Effect"
+import { pipe } from "../src/Function"
+import * as T from "../src/next/Effect"
 
 const prog = pipe(
   [0, 1, 2],
