+++ packages/core/tsconfig.json
@@ -6,5 +6,5 @@
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true
   },
-  "include": ["src/**/*", "scripts/**/*", "bench/**/*", "test/**/*"]
+  "include": ["src/**/*", "scripts/**/*", "bench/**/*", "test/**/*", "demo/**/*"]
 }
