+++ packages/server/README.md
@@ -1,1 +1,3 @@
 # WIP
+
+Native http server based on node's core http
